feat(ui): fold notifications surfaces into Org Settings (Phase 2c)
Retire the standalone NotificationsOfOrg monolith and redistribute its
surfaces to where they belong now that channels are catalog integrations:

- Subscriptions + Channel groups become Integrations sub-tab pills
  (Pro-gated), extracted into SubscriptionsOfOrg.vue / ChannelGroupsOfOrg.vue.
- Delivery History moves under a new Audit primary tab in Org Settings
  (Download Log + Delivery History sub-tabs), extracted into
  NotificationHistory.vue (Pro-gated sub-tab).
- The inbox becomes a nav-level icon + drawer with an unread badge in all
  editions, extracted into NotificationInbox.vue and mounted in TopNavBar
  (unread count polled on a 60s cadence at nav level).
- Quick Start wizard dropped entirely.
- Notifications left-nav item removed; /notificationsOfOrg/:orguuid now
  redirects to Integrations -> Subscriptions so old links resolve.

Shared stateless pieces (row interfaces, option arrays, tag/format/error
helpers, survived read-only queries, name-map factory) hoisted into
utils/notificationsCommon.ts so the extracted components stay in sync.

ReARM-Agent: 1420896f-adf5-4843-896f-d863cfcc6528
ReARM-Agentic-Session: 6fa3ea66-7d3d-4042-bca7-d5a75879a3e0
